third_party/WebKit/LayoutTests/http/tests/origin_trials/webexposed/animationworklet-origin-trial-interfaces-worklet-scope-expected.txt

CONSOLE MESSAGE: line 13: List all interfaces in the animation worklet global scope.
CONSOLE MESSAGE: line 147: interface AnimationWorkletGlobalScope : WorkletGlobalScope
CONSOLE MESSAGE: line 147:     method constructor
CONSOLE MESSAGE: line 147: interface ByteLengthQueuingStrategy
CONSOLE MESSAGE: line 147:     method constructor
CONSOLE MESSAGE: line 147:     method size
CONSOLE MESSAGE: line 147: interface CountQueuingStrategy
CONSOLE MESSAGE: line 147:     method constructor
CONSOLE MESSAGE: line 147:     method size
CONSOLE MESSAGE: line 147: interface DataView
CONSOLE MESSAGE: line 147:     getter buffer
CONSOLE MESSAGE: line 147:     getter byteLength
CONSOLE MESSAGE: line 147:     getter byteOffset
CONSOLE MESSAGE: line 147:     method constructor
CONSOLE MESSAGE: line 147:     method getFloat32
CONSOLE MESSAGE: line 147:     method getFloat64
CONSOLE MESSAGE: line 147:     method getInt16
CONSOLE MESSAGE: line 147:     method getInt32
CONSOLE MESSAGE: line 147:     method getInt8
CONSOLE MESSAGE: line 147:     method getUint16
CONSOLE MESSAGE: line 147:     method getUint32
CONSOLE MESSAGE: line 147:     method getUint8
CONSOLE MESSAGE: line 147:     method setFloat32
CONSOLE MESSAGE: line 147:     method setFloat64
CONSOLE MESSAGE: line 147:     method setInt16
CONSOLE MESSAGE: line 147:     method setInt32
CONSOLE MESSAGE: line 147:     method setInt8
CONSOLE MESSAGE: line 147:     method setUint16
CONSOLE MESSAGE: line 147:     method setUint32
CONSOLE MESSAGE: line 147:     method setUint8
CONSOLE MESSAGE: line 147: interface EffectProxy
CONSOLE MESSAGE: line 147:     getter localTime
CONSOLE MESSAGE: line 147:     method constructor
CONSOLE MESSAGE: line 147:     setter localTime
CONSOLE MESSAGE: line 147: interface ReadableStream
CONSOLE MESSAGE: line 147:     getter locked
CONSOLE MESSAGE: line 147:     method cancel
CONSOLE MESSAGE: line 147:     method constructor
CONSOLE MESSAGE: line 147:     method getReader
CONSOLE MESSAGE: line 147:     method pipeThrough
CONSOLE MESSAGE: line 147:     method pipeTo
CONSOLE MESSAGE: line 147:     method tee
CONSOLE MESSAGE: line 147: interface TransformStream
CONSOLE MESSAGE: line 147:     getter readable
CONSOLE MESSAGE: line 147:     getter writable
CONSOLE MESSAGE: line 147:     method constructor
CONSOLE MESSAGE: line 147: interface WorkletGlobalScope
CONSOLE MESSAGE: line 147:     method constructor
CONSOLE MESSAGE: line 147: interface WritableStream
CONSOLE MESSAGE: line 147:     getter locked
CONSOLE MESSAGE: line 147:     method abort
CONSOLE MESSAGE: line 147:     method constructor
CONSOLE MESSAGE: line 147:     method getWriter
CONSOLE MESSAGE: line 147: global object
CONSOLE MESSAGE: line 147:     attribute console
CONSOLE MESSAGE: line 147:     method gc
CONSOLE MESSAGE: line 147:     method registerAnimator
